+++ inst/application/_utilities/MDR/prepare_dehub_mdr.py
@@ -48,25 +48,39 @@ class PrepareMDR(MDRHandling):
   
   def prepare_mdr(self):
     # define names to keep,
-    keep_rows = [
-      "AdministrativesGeschlecht",
-      "Geburtsdatum"
-      ]
+    keep_rows = {
+      "AdministrativesGeschlecht": "Person.Demographie.AdministrativesGeschlecht",
+      "Geburtsdatum": "Person.Demographie.Geburtsdatum",
+      "PLZ": "Person.Demographie.Adresse.PLZ",
+      "Patienten-Identifikator": "Person.Patient.Patienten-Identifikator.Patienten-Identifikator",
+      "Aufnahmenummer": "Fall.Einrichtungskontakt.Aufnahmenummer",
+      "Beginndatum": "Fall.Einrichtungskontakt.Beginndatum",
+      "Enddatum": "Fall.Einrichtungskontakt.Enddatum",
+      "VollstaendigerDiagnosekode": "Diagnose.ICD10GMDiagnoseKodiert.VollstaendigerDiagnosekode"
+    }
     
     # duplicate rows for databases
     db_names = ["i2b2", "fhir_gw"]
     
+    # generate grid for every possible combination of designation, system_name and system_type
     pd_dat = [
       {
         "designation": x,
         "source_system_name": y,
-        "source_system_type": z} for x in keep_rows for y in db_names for z in ["postgres"]
+        "source_system_type": z} for x in keep_rows.keys() for y in db_names for z in ["postgres"]
       ]
     
     merge_df = pd.DataFrame(
       data = pd_dat
     )
     
+    # remove rows with missing definition
+    self.mdr.dropna(subset = ["definition"], inplace = True)
+    
+    # remove duplicate values
+    self.mdr.drop_duplicates(inplace=True)
+    
+    # drop columns that are being joined with merge_df
     self.mdr.drop(
       columns = ["source_system_name", "source_system_type"],
       inplace = True
@@ -77,16 +91,21 @@ class PrepareMDR(MDRHandling):
       how = "inner",
       on = "designation"
     )
-  
+    
+    # set some default values
     self.mdr.dqa_assessment = "1"
+    
+    # replace designation
+    self.mdr.replace({"designation": keep_rows}, inplace = True)
+    
+    # set keys and variable_name
     self.mdr.key = self.mdr.designation
-
-    self.mdr = self.mdr[self.mdr.designation.isin(keep_rows)]
+    self.mdr.variable_name = self.mdr.designation
     
 
 if __name__ == "__main__":
   prep = PrepareMDR(
-    mdr_file="dehub_mdr_clean.csv-20220207_144222.csv",
+    mdr_file="dehub_mdr_clean.csv-20220209_110106.csv",
     csv_separator="\t"
   )
   prep()